# backend/docs/swagger.yml
openapi: 3.0.0
info:
  title: GitParse API
  description: This is a documentation of GitParse API.
  version: 1.0.0
servers:
  - url: "https://gitparse"
paths:
  /projects:
    get:
      summary: List all projects (repositories)
      description: Get a list of all visible projects across Git for the authenticated user.
      parameters:
        # - name: platformName
        #   in: query
        #   required: true
        #   schema:
        #     type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 1
        - name: perPage
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30

      responses:
        '400':
          description: Invalid input
        '404':
          description: User not found
        '401':
          description: Unauthorized
        '200':
          description: A list of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
      security:
        - gitlab_auth:
            - write:users
              read:data
        - github_auth:
            - write:users
              read:data
  /projects/{repositoryId}:
    get:
      summary: Get a repository
      description: Get a repository by repositoryId
      parameters:
        # - name: resourceOwner
        #   in: path
        #   required: true
        #   schema:
        #     type: string
        - name: repositoryId
          in: path
          required: true
          schema:
            type: string
        # - name: platformName
        #   in: query
        #   required: true
        #   schema:
        #     type: string

      responses:
        '400':
          description: Invalid input
        '404':
          description: Repository not found
        '401':
          description: Unauthorized
        '200':
          description: Repository info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
      security:
        - gitlab_auth:
            - write:users
              read:data
        - github_auth:
            - write:users
              read:data
...
components:
  schemas:
...
    ProjectDto:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the project.
        description:
          type: string
          description: Description of the project.
        name:
          type: string
          description: The name of the project
        created_at:
          type: string
          format: date-time
          description: The date and time the project was created.
        default_branch:
          type: string
          description: The name of the main branch of the project.
        web_url:
          type: string
          format: uri
          description: The URL for viewing the project in the web interface.
        visibility:
          type: string
          description: The visibility level of the project (for example, public or private).
      required:
        - id
        - description
        - name
        - created_at
        - default_branch
        - web_url
        - visibility
  securitySchemes:
    gitlab_auth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://gitlab.com/oauth/authorize
          tokenUrl: https://gitlab.com/oauth/token
          scopes:
            write:users: edit user to repository
            read:data: read project info
    github_auth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://github.com/login/oauth/authorize
          tokenUrl: https://github.com/login/oauth/access_token
          scopes:
            write:users: edit user to repository
            read:data: read project info